refactor: pos-gmp.exo for prepareGMPPosition
